osmotr-auth-service-master/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
pydantic[email] 
passlib[bcrypt]           
python-multipart          
sqlalchemy
alembic
python-dotenv              
pydantic                   